% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labF.R
\name{labD}
\alias{labD}
\title{Labels all diagonal elements of symmetric matrices}
\usage{
labD(lab, i, sep = "_")
}
\arguments{
\item{lab}{prefix character}

\item{i}{number of rows/columns}
}
\value{
String vector arranged by column
}
\description{
Labels all diagonal elements of symmetric matrices
}
\examples{
library(OpenMx)

i = 5
mxMatrix("Diag", i, i, labels = labD("a", i))
}
