% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labF.R
\name{labF}
\alias{labF}
\title{Labels all elements of matrices}
\usage{
labF(lab, i, j, sep = "_")
}
\arguments{
\item{lab}{prefix character}

\item{i}{number of rows}

\item{j}{number of columns}
}
\value{
String vector arranged by column
}
\description{
Labels all elements of matrices
}
\examples{
library(OpenMx)

i = 5
j = 3
mxMatrix("Full", i, j, labels = labF("a", i, j))
}
