% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labF.R
\name{labL}
\alias{labL}
\title{Labels all elements of lower triangular matrices}
\usage{
labL(lab, i, sep = "_")
}
\arguments{
\item{lab}{prefix character}

\item{i}{number of rows/columns}
}
\value{
String vector arranged by column
}
\description{
Labels all elements of lower triangular matrices
}
\examples{
library(OpenMx)

i = 5
mxMatrix("Lower", i, i, labels = labL("a", i))
}
