% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labF.R
\name{labLL}
\alias{labLL}
\title{Labels all lower elements below the diaqonal of square matrices}
\usage{
labLL(lab, i, sep = "_")
}
\arguments{
\item{lab}{prefix character}

\item{i}{number of rows/columns}
}
\value{
String vector arranged by column
}
\description{
Labels all lower elements below the diaqonal of square matrices
}
\examples{
library(OpenMx)

i = 5
mxMatrix("Stand", i, i, labels = labL("a", i))
}
